reafactor: improved approach to finding new name
As names are used we already know which names can no longer be taken.
The new approach only lets new robots derive their name from the
available list.  This improves performance when large numbers
of robots already exist.  The previous approach could have a hard time
finding that last avaialable name (but in that case the robots are
taking over the world and we probably have more importan things to worry
about!).